+++ examples/adc-dma-rx.rs
@@ -9,29 +9,30 @@ use panic_halt as _;
 use cortex_m::{asm, singleton};
 
 use cortex_m_rt::entry;
-use stm32f1xx_hal::{adc, pac, prelude::*};
+use stm32f1xx_hal::{adc, pac, prelude::*, rcc};
 
 #[entry]
 fn main() -> ! {
     // Acquire peripherals
     let p = pac::Peripherals::take().unwrap();
     let mut flash = p.FLASH.constrain();
-    let rcc = p.RCC.constrain();
 
     // Configure ADC clocks
     // Default value is the slowest possible ADC clock: PCLK2 / 8. Meanwhile ADC
     // clock is configurable. So its frequency may be tweaked to meet certain
     // practical needs. User specified value is be approximated using supported
     // prescaler values 2/4/6/8.
-    let clocks = rcc.cfgr.adcclk(2.MHz()).freeze(&mut flash.acr);
+    let mut rcc = p
+        .RCC
+        .freeze(rcc::Config::hsi().adcclk(2.MHz()), &mut flash.acr);
 
-    let dma_ch1 = p.DMA1.split().1;
+    let dma_ch1 = p.DMA1.split(&mut rcc).1;
 
     // Setup ADC
-    let adc1 = adc::Adc::new(p.ADC1, &clocks);
+    let adc1 = adc::Adc::new(p.ADC1, &mut rcc);
 
     // Setup GPIOA
-    let mut gpioa = p.GPIOA.split();
+    let mut gpioa = p.GPIOA.split(&mut rcc);
 
     // Configure pa0 as an analog input
     let adc_ch0 = gpioa.pa0.into_analog(&mut gpioa.crl);