+++ examples/blinky_rtcalarm_irq.rs
@@ -78,10 +78,10 @@ fn main() -> ! {
     let dp = Peripherals::take().unwrap();
 
     let mut pwr = dp.PWR;
-    let rcc = dp.RCC.constrain();
+    let mut rcc = dp.RCC.constrain();
 
     // Set up the GPIO pin
-    let mut gpioc = dp.GPIOC.split();
+    let mut gpioc = dp.GPIOC.split(&mut rcc);
     let mut led = gpioc.pc13.into_push_pull_output(&mut gpioc.crh);
     let _ = led.set_high(); // Turn off
 
@@ -96,9 +96,9 @@ fn main() -> ! {
 
     // Set up the RTC
     // Enable writes to the backup domain
-    let mut backup_domain = rcc.bkp.constrain(dp.BKP, &mut pwr);
+    let mut backup_domain = dp.BKP.constrain(&mut pwr, &mut rcc);
     // Start the RTC
-    let mut rtc = Rtc::new(dp.RTC, &mut backup_domain);
+    let mut rtc = Rtc::new(dp.RTC, &mut backup_domain, &mut rcc);
     rtc.set_time(0);
     rtc.set_alarm(TOGGLE_INTERVAL_SECONDS);
     rtc.listen_alarm();