+++ examples/delay-timer-blinky.rs
@@ -11,7 +11,7 @@ use panic_halt as _; // panic handler
 use cortex_m_rt::entry;
 use stm32f1xx_hal as hal;
 
-use crate::hal::{pac, prelude::*};
+use crate::hal::{pac, prelude::*, rcc};
 
 #[entry]
 fn main() -> ! {
@@ -20,23 +20,19 @@ fn main() -> ! {
         cortex_m::peripheral::Peripherals::take(),
     ) {
         let mut flash = dp.FLASH.constrain();
-        // Set up the LED. On the BluePill it's connected to pin PC13.
-        let mut gpioc = dp.GPIOC.split();
-        let mut led = gpioc.pc13.into_push_pull_output(&mut gpioc.crh);
-
         // Set up the system clock. We want to run at 48MHz for this one.
-        let rcc = dp.RCC.constrain();
-        let clocks = rcc
-            .cfgr
-            .use_hse(8.MHz())
-            .sysclk(48.MHz())
-            .freeze(&mut flash.acr);
+        let mut rcc = dp
+            .RCC
+            .freeze(rcc::Config::hse(8.MHz()).sysclk(48.MHz()), &mut flash.acr);
+        // Set up the LED. On the BluePill it's connected to pin PC13.
+        let mut gpioc = dp.GPIOC.split(&mut rcc);
+        let mut led = gpioc.pc13.into_push_pull_output(&mut gpioc.crh);
 
         // Create a delay abstraction based on general-pupose 32-bit timer TIM2
 
-        //let mut delay = hal::timer::FTimerUs::new(dp.TIM2, &clocks).delay();
+        //let mut delay = hal::timer::FTimerUs::new(dp.TIM2, &mut rcc).delay();
         // or
-        let mut delay = dp.TIM2.delay_us(&clocks);
+        let mut delay = dp.TIM2.delay_us(&mut rcc);
 
         loop {
             // On for 1s, off for 3s.