+++ examples/usb_serial_rtic.rs
@@ -10,8 +10,11 @@ use panic_semihosting as _;
 #[rtic::app(device = stm32f1xx_hal::pac)]
 mod app {
     use cortex_m::asm::delay;
-    use stm32f1xx_hal::prelude::*;
-    use stm32f1xx_hal::usb::{Peripheral, UsbBus, UsbBusType};
+    use stm32f1xx_hal::{
+        prelude::*,
+        rcc,
+        usb::{Peripheral, UsbBus, UsbBusType},
+    };
     use usb_device::prelude::*;
 
     #[shared]
@@ -26,18 +29,14 @@ mod app {
     #[init(local = [usb_bus: Option<usb_device::bus::UsbBusAllocator<UsbBusType>> = None])]
     fn init(cx: init::Context) -> (Shared, Local, init::Monotonics) {
         let mut flash = cx.device.FLASH.constrain();
-        let rcc = cx.device.RCC.constrain();
+        let mut rcc = cx.device.RCC.freeze(
+            rcc::Config::hse(8.MHz()).sysclk(48.MHz()).pclk1(24.MHz()),
+            &mut flash.acr,
+        );
 
-        let clocks = rcc
-            .cfgr
-            .use_hse(8.MHz())
-            .sysclk(48.MHz())
-            .pclk1(24.MHz())
-            .freeze(&mut flash.acr);
+        assert!(rcc.clocks.usbclk_valid());
 
-        assert!(clocks.usbclk_valid());
-
-        let mut gpioa = cx.device.GPIOA.split();
+        let mut gpioa = cx.device.GPIOA.split(&mut rcc);
 
         // BluePill board has a pull-up resistor on the D+ line.
         // Pull the D+ pin down to send a RESET condition to the USB bus.
@@ -45,7 +44,7 @@ mod app {
         // will not reset your device when you upload new firmware.
         let mut usb_dp = gpioa.pa12.into_push_pull_output(&mut gpioa.crh);
         usb_dp.set_low();
-        delay(clocks.sysclk().raw() / 100);
+        delay(rcc.clocks.sysclk().raw() / 100);
 
         let usb_dm = gpioa.pa11;
         let usb_dp = usb_dp.into_floating_input(&mut gpioa.crh);