+++ src/gpio.rs
@@ -77,7 +77,7 @@ use core::convert::Infallible;
 use core::marker::PhantomData;
 
 use crate::afio;
-use crate::pac::EXTI;
+use crate::pac::{EXTI, RCC};
 
 mod partially_erased;
 pub use partially_erased::{PEPin, PartiallyErasedPin};
@@ -134,13 +134,13 @@ pub trait GpioExt {
     /// Splits the GPIO block into independent pins and registers.
     ///
     /// This resets the state of the GPIO block.
-    fn split(self) -> Self::Parts;
+    fn split(self, rcc: &mut RCC) -> Self::Parts;
 
     /// Splits the GPIO block into independent pins and registers without resetting its state.
     ///
     /// # Safety
     /// Make sure that all pins modes are set in reset state.
-    unsafe fn split_without_reset(self) -> Self::Parts;
+    unsafe fn split_without_reset(self, rcc: &mut RCC) -> Self::Parts;
 }
 
 /// Marker trait for active states.
@@ -395,10 +395,9 @@ macro_rules! gpio {
             impl GpioExt for $GPIOX {
                 type Parts = Parts;
 
-                fn split(self) -> Parts {
-                    let rcc = unsafe { &(*RCC::ptr()) };
-                    $GPIOX::enable(rcc);
-                    $GPIOX::reset(rcc);
+                fn split(self, rcc: &mut RCC) -> Parts {
+                        $GPIOX::enable(rcc);
+                        $GPIOX::reset(rcc);
 
                     Parts {
                         crl: Cr::<$port_id, false>,
@@ -409,9 +408,8 @@ macro_rules! gpio {
                     }
                 }
 
-                unsafe fn split_without_reset(self) -> Parts {
-                    let rcc = unsafe { &(*RCC::ptr()) };
-                    $GPIOX::enable(rcc);
+                unsafe fn split_without_reset(self, rcc: &mut RCC) -> Parts {
+                        $GPIOX::enable(rcc);
 
                     Parts {
                         crl: Cr::<$port_id, false>,