+++ src/timer/monotonic.rs
@@ -1,7 +1,7 @@
 //! RTIC Monotonic implementation
 
 use super::{FTimer, Instance};
-use crate::rcc::Clocks;
+use crate::rcc::Rcc;
 use core::ops::{Deref, DerefMut};
 pub use fugit::{self, ExtU32};
 use rtic_monotonic::Monotonic;
@@ -37,17 +37,17 @@ impl<TIM: Instance, const FREQ: u32> MonoTimer<TIM, FREQ> {
 }
 
 pub trait MonoTimerExt: Sized {
-    fn monotonic<const FREQ: u32>(self, clocks: &Clocks) -> MonoTimer<Self, FREQ>;
-    fn monotonic_us(self, clocks: &Clocks) -> MonoTimer<Self, 1_000_000> {
-        self.monotonic::<1_000_000>(clocks)
+    fn monotonic<const FREQ: u32>(self, rcc: &mut Rcc) -> MonoTimer<Self, FREQ>;
+    fn monotonic_us(self, rcc: &mut Rcc) -> MonoTimer<Self, 1_000_000> {
+        self.monotonic::<1_000_000>(rcc)
     }
 }
 
 macro_rules! mono {
     ($TIM:ty) => {
         impl MonoTimerExt for $TIM {
-            fn monotonic<const FREQ: u32>(self, clocks: &Clocks) -> MonoTimer<Self, FREQ> {
-                FTimer::new(self, clocks).monotonic()
+            fn monotonic<const FREQ: u32>(self, rcc: &mut Rcc) -> MonoTimer<Self, FREQ> {
+                FTimer::new(self, rcc).monotonic()
             }
         }