+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   Add `RInto` trait and `Rmp` peripheral wrapper, add `remap` for peripherals. [#514] [#520]
   Remove `RemapStruct`s. [#462] [#506] [#509]
 - Use independent `Spi` and `SpiSlave` structures instead of `OP` generic [#462]
-- Take `&Clocks` instead of `Clocks` [#498]
+- Include `Clocks` in `Rcc`. Take `&mut Rcc/RCC` where possible [#498] [#536]
 - Update to `stm32f1` v0.16.0 [#503] [#534]
 - `Spi` now takes `Option<PIN>` for `SCK`, `MISO`, `MOSI` [#514],
   add `SPIx::NoSck`, etc. [#537]
@@ -77,6 +77,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [#526]: https://github.com/stm32-rs/stm32f1xx-hal/pull/526
 [#528]: https://github.com/stm32-rs/stm32f1xx-hal/pull/528
 [#534]: https://github.com/stm32-rs/stm32f1xx-hal/pull/534
+[#536]: https://github.com/stm32-rs/stm32f1xx-hal/pull/536
 
 ## [v0.10.0] - 2022-12-12